backend/requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-ninja==1.5.0
django-simple-history==3.10.1
orjson==3.11.4
psycopg==3.1.18
psycopg-binary==3.2.12
pydantic==2.12.4
pydantic_core==2.41.5
python-decouple==3.8
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
